refactor: restructure CLI/GUI architecture following MVC pattern
- Extract GuiApplication from Main class for symmetry with CliApplication
- Refactor Main to be a pure orchestrator (persistence config + interface launcher)
- Improve CliLoginGraphicController with Optional, constants, and better error handling
- Create CliMainMenuGraphicController for main menu logic (Login/SignUp/Exit)
- Refactor CliView to remove direct System.out usage via CliViewFactory
- Move both GuiApplication and CliApplication to view package
- Ensure single shared LoginController for both GUI and CLI (MVC compliance)

+++ src/main/java/it/uniroma2/hoophub/dao/csv/NotificationDaoCsv.java
@@ -422,7 +422,7 @@ public class NotificationDaoCsv extends AbstractCsvDao implements NotificationDA
     public synchronized boolean deleteById(Long id) throws DAOException {
         validatePositiveId(id);
 
-        boolean found = deleteById(id.longValue(), COL_ID);
+        boolean found = deleteById(id, COL_ID);
 
         if (found) {
             logger.log(Level.INFO, "Notification deleted successfully: {0}", id);
